Split structs.go
